+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/archived23_24SeaonCenterStage/yiseArchived/IntakeSystem.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.archived23_24SeaonCenterStage.yise;
+package org.firstinspires.ftc.teamcode.archived23_24SeaonCenterStage.yiseArchived;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
